# Conflicts:
#	Record/Record/Base.lproj/Main.storyboard
#	Record/Record/HomePage/View/HomeTableViewCell.m

+++ Record/Record/HomePage/Model/HomeContentModel.h
@@ -14,9 +14,9 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic,copy)NSString *objectId;//id
 @property (nonatomic,copy)NSString *mobile;//加*手机号
 @property (nonatomic,copy)NSString *content;//内容
-@property (nonatomic,assign)int type;//类别   开心，愤怒，伤心，平常心
+@property (nonatomic,assign)NSInteger type;//类别   开心，愤怒，伤心，平常心
 @property (nonatomic,copy)NSString *date; //时间
-@property (nonatomic,assign)int praise;//点赞数
+@property (nonatomic,assign)NSInteger praise;//点赞数
 @end
 
 NS_ASSUME_NONNULL_END
